[Refactor] Move constants to Constants interface
