--- src/main/java/frc/robot/Config.java
@@ -1,4 +1,0 @@
-package frc.robot;
-
-public final class Config {
-}
