+++ src/main/java/frc/robot/subsystems/Elevator.java
@@ -4,16 +4,14 @@ import com.revrobotics.CANSparkMax;
 import com.revrobotics.CANSparkMaxLowLevel;
 
 import edu.wpi.first.wpilibj.command.Subsystem;
+import frc.robot.Constants;
 import frc.robot.commands.elevator.MoveWithJoystick;
 import frc.util.DriveHelper;
 
 
-public final class Elevator extends Subsystem {
+public final class Elevator extends Subsystem implements Constants {
   private static final double THROTTLE_DEADBAND = 0.05;
   private static final double HEADING_DEADBAND = 0.05;
-  
-  private final int ELEVATOR_MASTER = 8, // 6
-  ELEVATOR_SLAVE = 9;
 
   private CANSparkMax
 		elevatorMaster = new CANSparkMax(ELEVATOR_MASTER, CANSparkMaxLowLevel.MotorType.kBrushless),
