+++ src/main/java/frc/robot/subsystems/Intake.java
@@ -3,13 +3,10 @@ package frc.robot.subsystems;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 
 import edu.wpi.first.wpilibj.command.Subsystem;
+import frc.robot.Constants;
 import frc.robot.commands.intake.TiltArm;
 
-public class Intake extends Subsystem {
-
-  private final static int 
-    INTAKE_MOTOR = 8,
-    TILT_MOTOR = 9;
+public class Intake extends Subsystem implements Constants {
 
   public static final WPI_TalonSRX
     intakeMotor = new WPI_TalonSRX(INTAKE_MOTOR),
